+++ src/app/app-routing.module.ts
@@ -10,7 +10,7 @@ const routes: Routes = [
     },
     {
         path: 'home',
-        loadChildren: () => import('./home/home.module').then((m) => m.HomePageModule),
+        loadChildren: () => import('./modules/home/home.module').then((m) => m.HomePageModule),
     },
     {
         path: '',
@@ -29,4 +29,4 @@ const routes: Routes = [
     ],
     exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
