+++ src/app/modules/home/components/card-grid-item/card-grid-item.component.scss
