+++ src/app/modules/home/home.module.ts
@@ -7,11 +7,11 @@ import { NgxTippyModule } from 'ngx-tippy-wrapper';
 import { NgxVisibilityModule } from 'ngx-visibility';
 import { TranslocoModule } from '@ngneat/transloco';
 
-import { CardGridComponent } from '@app/home/components/card-grid/card-grid.component';
-import { CardGridItemComponent } from '@app/home/components/card-grid-item/card-grid-item.component';
-import { HomePage } from '@app/home/home.page';
-import { HomePageRoutingModule } from '@app/home/home-routing.module';
-import { HomeStateModule } from '@app/home/store/home-state.module';
+import { CardGridComponent } from '@app/modules/home/components/card-grid/card-grid.component';
+import { CardGridItemComponent } from '@app/modules/home/components/card-grid-item/card-grid-item.component';
+import { HomePage } from '@app/modules/home/home.page';
+import { HomePageRoutingModule } from '@app/modules/home/home-routing.module';
+import { HomeStateModule } from '@app/modules/home/store/home-state.module';
 import { ImageCardModule } from '@app/shared/components/image-card/image-card.module';
 import { SkeletonBlockModule } from '@app/shared/components/skeleton-block/skeleton-block.module';
 import { TypedTemplateDirective } from '@app/shared/directives/typed-template.directive';
