+++ src/app/modules/home/store/anime-rates/helpers/anime-rates-store-key.helpers.ts
@@ -2,7 +2,7 @@ import {
     LoadStatusKeysType,
     StatusKeysType,
     StatusPageType,
-} from '@app/home/store/anime-rates/types/anime-rates-store.interface';
+} from '@app/modules/home/store/anime-rates/types/anime-rates-store.interface';
 import { UserRateStatusType } from '@app/shared/types/shikimori/user-rate-status.type';
 
 export function getRateLoadedKey(status: UserRateStatusType): LoadStatusKeysType {
