+++ src/app/modules/home/store/anime-rates/helpers/anime-rates.helpers.ts
@@ -1,11 +1,10 @@
 import { AnimeNameSortingConfig, sortByAnimeUserRating } from '@app/shared/utils/rx-anime-rates-functions';
 import { UserAnimeRate } from '@app/shared/types/shikimori/user-anime-rate';
-import { UserRateStatusType } from '@app/shared/types/shikimori/user-rate-status.type';
 
 export function isDuplicateArrayFilter(item: UserAnimeRate, pos: number, array: UserAnimeRate[]) {
     return !pos || item.anime.id !== array[pos - 1].anime.id;
 }
 
-export function sortAnimeRatesByUserRating(status: UserRateStatusType, animeNameSortCfg?: AnimeNameSortingConfig) {
+export function sortAnimeRatesByUserRating(animeNameSortCfg?: AnimeNameSortingConfig) {
     return (a: UserAnimeRate, b: UserAnimeRate) => sortByAnimeUserRating(a, b, animeNameSortCfg);
 }
