+++ src/app/modules/home/store/anime-rates/reducers/anime-rates.reducer.ts
@@ -1,16 +1,16 @@
 import { createReducer, on } from '@ngrx/store';
 
-import AnimeRatesStoreInterface from '@app/home/store/anime-rates/types/anime-rates-store.interface';
+import AnimeRatesStoreInterface from '@app/modules/home/store/anime-rates/types/anime-rates-store.interface';
 import {
     allPagesLoadedForStatusAction,
     incrementPageForStatusAction,
-} from '@app/home/store/anime-rates/actions/anime-rate-paging.actions';
+} from '@app/modules/home/store/anime-rates/actions/anime-rate-paging.actions';
 import {
     getRateLoadedKey,
     getRatePageKey,
     getRateStoreKey,
-} from '@app/home/store/anime-rates/helpers/anime-rates-store-key.helpers';
-import { loadAnimeRateByStatusSuccessAction } from '@app/home/store/anime-rates/actions/load-anime-rate.action';
+} from '@app/modules/home/store/anime-rates/helpers/anime-rates-store-key.helpers';
+import { loadAnimeRateByStatusSuccessAction } from '@app/modules/home/store/anime-rates/actions/load-anime-rate.action';
 
 const initialState: AnimeRatesStoreInterface = {
     planned: [],
