+++ src/app/modules/home/store/anime-rates/types/anime-rates-store.interface.ts
