+++ src/app/modules/home/types/anime-grid.interface.ts
