+++ src/app/modules/home/types/visibility-change.interface.ts
